# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3

info:
  title: Forum API
  description: REST API for forum posts and users (JWT bearer auth).
  version: 1.0.0

servers:
  - url: http://localhost:9000
    description: Local

tags:
  - name: Auth
    description: Authentication
  - name: Posts
    description: Post operations
  - name: Replies
    description: Replies on posts
  - name: Users
    description: User operations
  - name: Theme
    description: Background customization per page / per user

paths:

  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Obtain a JWT (username + password)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials

  /auth/register:
    post:
      tags: [Auth]
      operationId: register
      summary: Self-service signup; creates a USER account and returns a JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "409":
          description: Username or email already taken

  /posts:
    get:
      tags: [Posts]
      operationId: listPosts
      summary: List topics, newest first, paginated 15 per page (FUS1)
      security: []
      parameters:
        - name: page
          in: query
          required: false
          description: Zero-based page index.
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          description: Page size.
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 15
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostPage"

    post:
      tags: [Posts]
      operationId: createPost
      summary: Create a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostResponse"

  /posts/{id}:
    get:
      tags: [Posts]
      operationId: getPostById
      summary: Get post by id
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostResponse"
        "404":
          description: Not found

    put:
      tags: [Posts]
      operationId: updatePost
      summary: Update a topic (author, moderator, or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePostRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostResponse"
        "403":
          description: Not allowed
        "404":
          description: Not found
        "409":
          description: Title already in use

    delete:
      tags: [Posts]
      operationId: deletePost
      summary: Delete a topic (author, moderator, or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: No content
        "403":
          description: Not allowed
        "404":
          description: Not found

  /posts/{id}/likes:
    post:
      tags: [Posts]
      operationId: likePost
      summary: Like a topic (idempotent)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostResponse"
        "404":
          description: Not found

    delete:
      tags: [Posts]
      operationId: unlikePost
      summary: Remove a like from a topic (idempotent)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostResponse"
        "404":
          description: Not found

  /posts/{postId}/replies:
    get:
      tags: [Replies]
      operationId: listRepliesForPost
      summary: List replies for a post, paginated 10 per page (FUS3)
      security: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          required: false
          description: Zero-based page index.
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          description: Page size.
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplyPage"
        "404":
          description: Post not found

    post:
      tags: [Replies]
      operationId: createReply
      summary: Create a reply on a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReplyRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplyResponse"
        "404":
          description: Post not found

  /replies/{id}:
    get:
      tags: [Replies]
      operationId: getReplyById
      summary: Get reply by id
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplyResponse"
        "404":
          description: Not found

    put:
      tags: [Replies]
      operationId: updateReply
      summary: Update a reply (author, moderator, or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReplyRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplyResponse"
        "403":
          description: Not allowed
        "404":
          description: Not found

    delete:
      tags: [Replies]
      operationId: deleteReply
      summary: Delete a reply (author, moderator, or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: No content
        "403":
          description: Not allowed
        "404":
          description: Not found

  /replies/{id}/likes:
    post:
      tags: [Replies]
      operationId: likeReply
      summary: Like a reply (idempotent)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplyResponse"
        "404":
          description: Not found

    delete:
      tags: [Replies]
      operationId: unlikeReply
      summary: Remove a like from a reply (idempotent)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplyResponse"
        "404":
          description: Not found

  /users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: List all users
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponse"

    post:
      tags: [Users]
      operationId: createUser
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "409":
          description: Username or email already in use

  /users/{id}:
    get:
      tags: [Users]
      operationId: getUserById
      summary: Get user by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          description: Not found

    put:
      tags: [Users]
      operationId: updateUser
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          description: Not found
        "409":
          description: Username or email already in use

    delete:
      tags: [Users]
      operationId: deleteUser
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: No content
        "404":
          description: Not found

  /theme-customizations/home:
    get:
      tags: [Theme]
      operationId: getHomeTheme
      summary: Background customization for the main page
      security: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThemeCustomization"
    put:
      tags: [Theme]
      operationId: updateHomeTheme
      summary: Update the main page theme (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateThemeRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThemeCustomization"

  /theme-customizations/users/{userId}:
    get:
      tags: [Theme]
      operationId: getUserTheme
      summary: Background customization for a user's content (defaults if unset)
      security: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThemeCustomization"
    put:
      tags: [Theme]
      operationId: updateUserTheme
      summary: Update a user's theme (self or admin)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateThemeRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThemeCustomization"

components:

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 100
        password:
          type: string
          minLength: 1
          maxLength: 128

    RegisterRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 320
        password:
          type: string
          minLength: 8
          maxLength: 72
          description: Stored as a bcrypt hash; max 72 bytes for bcrypt.

    LoginResponse:
      type: object
      required: [accessToken, tokenType, expiresInSeconds]
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresInSeconds:
          type: integer
          format: int64

    UserRole:
      type: string
      description: USER denotes a regular (non-privileged) account.
      enum: [ADMIN, MODERATOR, USER]

    CreateUserRequest:
      type: object
      required: [username, role, password]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 320
        role:
          $ref: "#/components/schemas/UserRole"
        password:
          type: string
          minLength: 8
          maxLength: 72
          description: Stored as a bcrypt hash; max 72 bytes for bcrypt.

    UpdateUserRequest:
      type: object
      required: [username, role]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 320
          nullable: true
        role:
          $ref: "#/components/schemas/UserRole"
        password:
          type: string
          minLength: 8
          maxLength: 72
          description: If set, replaces the password.

    UserResponse:
      type: object
      required: [id, username, role, createdAt]
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
          format: email
          nullable: true
        role:
          $ref: "#/components/schemas/UserRole"
        createdAt:
          type: string
          format: date-time

    ThemeCustomization:
      type: object
      description: >-
        Drives the client-rendered doodle background. `seed` (a UUID)
        deterministically controls icon placement, rotation, size and shade;
        `figures` are Font Awesome names.
      required:
        [scope, seed, figures, bgColor, bgOpacity, cardColor, cardOpacity,
         opacity, blur, density, baseSize, sizeVariation, accentColor,
         accentVariation]
      properties:
        scope:
          type: string
          enum: [PAGE, USER]
        pageKey:
          type: string
          nullable: true
          description: Set for PAGE scope, e.g. "home".
        userId:
          type: integer
          format: int64
          nullable: true
          description: Set for USER scope.
        seed:
          type: string
          description: UUID seeding all deterministic randomness.
        figures:
          type: array
          items:
            type: string
          description: Font Awesome icon names, e.g. ["gift", "star", "heart"].
        bgColor:
          type: string
          description: CSS color for the page background, e.g. "#ffffff".
        bgOpacity:
          type: number
          format: float
          description: Opacity (0..1) of the background colour fill.
        cardColor:
          type: string
          description: Background color of post/reply/list cards.
        cardOpacity:
          type: number
          format: float
          description: Opacity (0..1) of the card background.
        opacity:
          type: number
          format: float
          description: Icon opacity (0..1) — strength of the pattern.
        blur:
          type: integer
          description: Blur applied to the icons, in pixels.
        density:
          type: integer
          description: Number of icons placed in the background.
        baseSize:
          type: integer
          description: General icon size in pixels.
        sizeVariation:
          type: number
          format: float
          description: 0..1 — how often an icon is rendered notably larger.
        accentColor:
          type: string
          description: Base color the icons are tinted with, e.g. "#f97316".
        accentVariation:
          type: number
          format: float
          description: 0..1 — fraction of icons rendered as darker shades.

    UpdateThemeRequest:
      type: object
      description: Editable theme fields (scope/keys are fixed by the path).
      required:
        [seed, figures, bgColor, bgOpacity, cardColor, cardOpacity, opacity,
         blur, density, baseSize, sizeVariation, accentColor, accentVariation]
      properties:
        seed:
          type: string
        figures:
          type: array
          items:
            type: string
        bgColor:
          type: string
        bgOpacity:
          type: number
          format: float
        cardColor:
          type: string
        cardOpacity:
          type: number
          format: float
        opacity:
          type: number
          format: float
        blur:
          type: integer
        density:
          type: integer
        baseSize:
          type: integer
        sizeVariation:
          type: number
          format: float
        accentColor:
          type: string
        accentVariation:
          type: number
          format: float

    CreatePostRequest:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
          maxLength: 500
        content:
          type: string
          maxLength: 10000
    AuthorSummary:
      type: object
      description: Minimal info about the user who created an object.
      required: [id, username, role]
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"

    UpdatePostRequest:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
          maxLength: 500
        content:
          type: string
          maxLength: 10000

    PostResponse:
      type: object
      required: [id, title, content, author, views, likeCount, likedByCurrentUser, createdAt, updatedAt]
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        content:
          type: string
        author:
          $ref: "#/components/schemas/AuthorSummary"
        views:
          type: integer
          format: int64
          description: Number of times the topic has been opened.
        likeCount:
          type: integer
          format: int64
        likedByCurrentUser:
          type: boolean
          description: Whether the requesting user has liked this topic (false when anonymous).
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateReplyRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          maxLength: 10000

    UpdateReplyRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          maxLength: 10000

    ReplyResponse:
      type: object
      required: [id, postId, content, author, likeCount, likedByCurrentUser, createdAt, updatedAt]
      properties:
        id:
          type: integer
          format: int64
        postId:
          type: integer
          format: int64
        content:
          type: string
        author:
          $ref: "#/components/schemas/AuthorSummary"
        likeCount:
          type: integer
          format: int64
        likedByCurrentUser:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    PostPage:
      type: object
      required: [content, page, size, totalElements, totalPages]
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/PostResponse"
        page:
          type: integer
          format: int32
          description: Zero-based page index.
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32

    ReplyPage:
      type: object
      required: [content, page, size, totalElements, totalPages]
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/ReplyResponse"
        page:
          type: integer
          format: int32
          description: Zero-based page index.
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        newest:
          description: >-
            The single newest reply for the topic, for pinning at the top
            regardless of the current page. Null when there are no replies.
          allOf:
            - $ref: "#/components/schemas/ReplyResponse"

security:
  - bearerAuth: []